Add posts controller
