+++ app.py
@@ -1,62 +1,20 @@
 import os
 from flask_cors import cross_origin
 from api import app
-from api.util.decorators import token_required
-from api.services.posts import Filtros, Categorias, ListaPostagens, Postagens, PostagensId, Recomendados, Selo
-from api.controller import users, auth, forms, notifications, comments
+from api.controller import users, auth, forms, notifications, comments, posts
 
 app.register_blueprint(users.app)
 app.register_blueprint(auth.app)
 app.register_blueprint(forms.app)
 app.register_blueprint(notifications.app)
 app.register_blueprint(comments.app)
+app.register_blueprint(posts.app)
 
+#TODO: Mover para controller
 @app.route('/')
 @cross_origin(origin='*', headers=['Content-Type', 'Authorization'])
 def hello():
 	return "This API Works! [" + os.environ.get("ENV", "DEV") + "]"
-
-@app.route('/categorias', methods=['POST', 'GET'])
-@cross_origin(origin='*', headers=['Content-Type', 'Authorization'])
-@token_required
-def categorias():
-    return Categorias()
-
-@app.route('/selo/<id>', methods=['PUT'])
-@cross_origin(origin='*', headers=['Content-Type', 'Authorization'])
-@token_required
-def selo(id):
-    return Selo(id)
-
-@app.route('/postagens', methods=['POST', 'GET'])
-@cross_origin(origin='*', headers=['Content-Type', 'Authorization'])
-@token_required
-def postagens():
-    return Postagens()
-
-@app.route('/recomendados', methods=['GET'])
-@cross_origin(origin='*', headers=['Content-Type', 'Authorization'])
-@token_required
-def recomendados():
-    return Recomendados()
-
-@app.route('/postagens/categorias/<id_categoria>', methods=['GET'])
-@cross_origin(origin='*', headers=['Content-Type', 'Authorization'])
-@token_required
-def filtros(id_categoria):
-    return Filtros(id_categoria)
-
-@app.route('/postagens/<id>', methods=['GET'])
-@cross_origin(origin='*', headers=['Content-Type', 'Authorization'])
-@token_required
-def postagensId(id):
-    return PostagensId(id)
-
-@app.route('/lista_postagens/<id>', methods=['GET'])
-@cross_origin(origin='*', headers=['Content-Type', 'Authorization'])
-@token_required
-def lista_postagens(id):
-    return ListaPostagens(id)
 
 if __name__ == '__main__':
     app.run(host="0.0.0.0", port=8000)
